initialize new updater.
